+++ trunk/src/com/geoscope/GeoLog/DEVICE/VideoRecorderModule/SpyDroid/CameraStreamerFRAME.java
@@ -494,7 +494,7 @@ public class CameraStreamerFRAME extends Camera {
 	 
 	@SuppressWarnings("deprecation")
 	@Override
-	public void Setup(SurfaceHolder holder, String ip, int audio_port, int video_port, int Mode, int asrc, int sps, int abr, int vsrc, int resX, int resY, int fps, int br, int UserID, String UserPassword, int pidGeographServerObject, boolean pflTransmitting, boolean pflSaving, boolean pflAudio, boolean pflVideo, boolean pflPreview, double MaxMeasurementDuration) throws Exception {
+	public void Setup(SurfaceHolder holder, String ip, int audio_port, int video_port, int Mode, int asrc, int sps, int abr, int vsrc, int resX, int resY, int fps, int br, long UserID, String UserPassword, long pidGeographServerObject, boolean pflTransmitting, boolean pflSaving, boolean pflAudio, boolean pflVideo, boolean pflPreview, double MaxMeasurementDuration) throws Exception {
 		flAudio = pflAudio;
 		flVideo = pflVideo;
 		flTransmitting = pflTransmitting;
@@ -750,7 +750,7 @@ public class CameraStreamerFRAME extends Camera {
 	}
 	
 	@Override
-	public void StartTransmitting(int pidGeographServerObject) {
+	public void StartTransmitting(long pidGeographServerObject) {
 		//.
 		flTransmitting = true;
 	}
